+++ CHANGELOG.md
@@ -7,3 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+
+## 0.0.2
+- Fixed bug fith get top scripts. Getting up to 100 scripts
+- Implemented Custom Calculations